new updated system
error with delays and controller. reads wrong values in fibonacci. check
delays.
controller_en = when cache is done reading or writing.

controller waits until the cache is done reading.
first delay waits until atleast 3 cycles to latch the read signal to the
cache. this is the minimum clock cycles that the Read or write need to
be enabled.

try changing the number of cycles in  numcyclesDelay